+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 - Add auto-refresh functionality to the Log page (#957, @mickBoat00)
 - Redesign the "Status Badges" page
 - Add support for per-check status badges (#853)
+- Add "Last ping subject" field in email notifications
 
 ### Bug Fixes
 - Fix Gotify integration to handle Gotify server URLs with paths (#964)